perf(hash): generate hash types in macro
