+++ mmr/bridge/header.rs
@@ -1,84 +1,6 @@
-//! format bridge
-#![macro_use]
+use super::{Bloom, U256};
 use scale::{Decode, Encode};
-use std::fmt::{Debug, Display, Formatter, Result as FmtResult};
-use uint::construct_uint;
-
-construct_uint! {
-    #[derive(Encode, Decode)]
-    pub struct U256(4);
-}
-
-#[macro_export]
-/// Convert bytes to hex
-macro_rules! hex {
-    ($bytes:expr) => {{
-        let mut s = String::new();
-        for i in $bytes {
-            s.push_str(&format!("{:02x}", i));
-        }
-        s
-    }};
-}
-
-#[macro_export]
-/// Convert hext to `Vec<u8>` or `[u8; n]`
-macro_rules! bytes {
-    // Convert hex to Vec<u8>
-    ($hex:expr) => {{
-        let mut h = $hex;
-        if h.starts_with("0x") {
-            h = &h[2..];
-        }
-
-        (0..h.len())
-            .step_by(2)
-            .map(|i| u8::from_str_radix(&h[i..i + 2], 16))
-            .collect::<Result<Vec<u8>, _>>()
-            .unwrap_or_default()
-    }};
-
-    // Convert hex to [u8; $bits]
-    ($hex:expr, $bits:expr) => {{
-        let mut hash = [0_u8; $bits];
-        hash.copy_from_slice(&bytes!($hex));
-        hash
-    }};
-}
-
-#[derive(Decode, Encode)]
-struct Bloom(pub [u8; 256]);
-
-impl Display for Bloom {
-    fn fmt(&self, f: &mut Formatter<'_>) -> FmtResult {
-        f.write_str(&hex!(self.0.as_ref()))
-    }
-}
-
-impl Debug for Bloom {
-    fn fmt(&self, f: &mut Formatter<'_>) -> FmtResult {
-        f.debug_list().entries(self.0.iter()).finish()
-    }
-}
-
-impl Default for Bloom {
-    fn default() -> Bloom {
-        Bloom([0; 256])
-    }
-}
-
-impl PartialEq for Bloom {
-    fn eq(&self, other: &Self) -> bool {
-        for i in 0..self.0.len() {
-            if self.0[i] != other.0[i] {
-                return false;
-            }
-        }
-        true
-    }
-}
-
-impl Eq for Bloom {}
+use std::fmt::Debug;
 
 /// Darwinia Eth header
 #[derive(Decode, Encode, Debug, PartialEq, Eq)]
